+++ CHANGELOG.md
@@ -7,3 +7,9 @@
 ## [1.0.1]
 
 - Update README & CHANGELOG
+
+## [2.0.0]
+
+- Add more function:
+1. Right-click the markdown file in explorer
+2. Auto run command if the project hasn't been run